requirements.txt
flask==3.0.0
google-cloud-storage==2.14.0
sendgrid==6.11.0
jinja2==3.1.3
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0